+++ modulo2/projeto-astromatch/astromatch/src/App.js
@@ -7,46 +7,28 @@ import Home from "./pages/Home/Home.js"
 import MyMatches from "./pages/MyMatches/MyMatches.js"
 
 function App() {
-  const [profileToChoose, setProfileToChoose] = useState({});
   const [matchesList, setMatchesList] = useState([]);
   const [screen, setScreen] = useState("home");
 
   const changeScreen = () => {
     switch (screen) {
       case "home":
-        return (<Home/>);
-      case "my matches":
+        return (
+          <Home
+          />
+        );
+      case "myMatches":
         return (<MyMatches />);
       default:
         return (<Home/>);
     }
   };
 
-  useEffect(() => {
-    getProfileToChoose();
-  }, []);
-
-  const getProfileToChoose = async () => {
-    try {
-      const response = await axios.get(`${baseUrl}/person`);
-      setProfileToChoose(response.data.profile);
-      // console.log("profile", profileToChoose);
-    } catch (err) {
-      console.log(err);
-    };
-  };
-
-  const postChoosePerson = async (currentId, userChoice) => {
-    const body = {
-      id: currentId,
-      choice: userChoice
-    }
-    try {
-      const response = await axios.post(`${baseUrl}/choose-person`, body, headersConfig);
-      getProfileToChoose();
-      // console.log(response);
-    } catch (err) {
-      console.log(err);
+  const onClickChangeScreen = () => {
+    if (screen === "home") {
+      setScreen("myMatches");
+    } else {
+      setScreen("Home");
     };
   };
 
@@ -65,17 +47,11 @@ function App() {
       {/* <button onClick={() => getMatches()}>Ver meus matches</button> */}
       <Header 
         screen={screen}
+        onClickChangeScreen={() => onClickChangeScreen()}
       />
 
-      <div>
-        <h3>{profileToChoose.name}, {profileToChoose.age}</h3>
-        <p>{profileToChoose.bio}</p>
-        <img src={profileToChoose.photo} alt="Foto de perfil" />
-      </div>
-      
-      <button onClick={() => postChoosePerson(profileToChoose.id, false)}>Não Quero</button>
-      <button onClick={() => postChoosePerson(profileToChoose.id, true)}>Quero</button>
-      
+      {changeScreen()}
+
     </div>
   );
 }
