+++ modulo2/projeto-astromatch/astromatch/src/components/header/Header.js
@@ -3,7 +3,7 @@ import fireHeart from "./../../images/heart.png"
 import searchHeart from "./../../images/search.png"
 import styled from "styled-components";
 
-const Head = styled.header`
+const HeaderPages = styled.header`
     display: flex;
     justify-content: space-around;
     align-items: center;
@@ -20,17 +20,17 @@ const Button = styled.button`
 `
 export default function Header(props) {
     return (
-        <Head>
+        <HeaderPages>
             <Logo src={logo} alt="Ícone planeta de coração"/>
             <h1>AstroMatch</h1>
             {props.screen === "home"
-            ? (<Button title="Ver meus matches">
+            ? (<Button title="Ver meus matches" onClick={props.onClickChangeScreen} >
                     <img src={fireHeart} alt="Ícone de coração em chamas" />
                </Button>)
-            : (<Button title="Voltar para os perfis">
+            : (<Button title="Voltar para os perfis" onClick={props.onClickChangeScreen} >
                     <img src={searchHeart} alt="Íconde de lupa com coração" />
                </Button>
             )}
-        </Head>
+        </HeaderPages>
     );
 };
